+++ lib/src/meta/owner_reference.g.dart
@@ -8,12 +8,12 @@ part of 'owner_reference.dart';
 
 OwnerReference _$OwnerReferenceFromJson(Map<String, dynamic> json) =>
     OwnerReference()
-      ..apiVersion = json['apiVersion'] as String
-      ..blockOwnerDeletion = json['blockOwnerDeletion'] as bool
-      ..controller = json['controller'] as bool
-      ..kind = json['kind'] as String
-      ..name = json['name'] as String
-      ..uid = json['uid'] as String;
+      ..apiVersion = json['apiVersion'] as String?
+      ..blockOwnerDeletion = json['blockOwnerDeletion'] as bool?
+      ..controller = json['controller'] as bool?
+      ..kind = json['kind'] as String?
+      ..name = json['name'] as String?
+      ..uid = json['uid'] as String?;
 
 Map<String, dynamic> _$OwnerReferenceToJson(OwnerReference instance) =>
     <String, dynamic>{